Fix issue where ReadExifOrientation() threw DirectoryNotFoundException
Originally, the user's selected image was passed to ReadExifOrientation() via uri (string). ExifReader's constructor implements FileStream's constructor.
Some kind of access error must have occurred; in any case, the fix was to pass the selected image via memory instead (byte[] array). The UnityWebRequest class
was used to retrieve the reference to the image's bytes, which were then passed to ExifReader's overloaded constructor.

+++ Assets/#Scripts/PolaroidManager.cs
@@ -34,7 +34,7 @@ public class PolaroidManager : MonoBehaviour
         currPolaroid = null;
     }
 
-    public void RotatePicFrame(int orientationValue)
+    public void RotatePicFrame(short orientationValue)
     {
         if (orientationValue == 0)
             return;
